Refactor helper into init_sdk, login_helper, and other helpers.
Iterating and refining login_helper is the primary goal.
Currently it still assumes that an 'sdk' global is defined,
which should probably be fixed.

Also I've done a manual wrapping to isolate the implementation,
but I imagine that we should have automated tooling to generate
the appropriate translation and wrapping for use in various
contexts. Along those lines, I've used 'let' to match the style
in the rest of the code base, but without translation that
would limit our browser compatibility, right?

I also wonder whether getFluxCredentials and clearFluxCredentials
is actually necessary. It seems like the correct credentials
could just be the resolving value of the checkLogin() promise.
We should be able to handle the situation where the stored
credentials have expired, however.

We should also still improve the random number generation to an
appropriate standard.