Added a console to display notifications
Added a new notification console that displays any kind of notification
event dispatched by the selected project’s data table.

This required a few changes to the way notifications are hooked. The
handler is moved from helper.js to index.js so it is possible to write
general events to the console.

Note that this commit introduces one known issue: The web socket
handlers are not removed when the project changes.