Fixes auth redirection
Moving the source code around (i.e., into /src) affected
the URLs, since the main tutorial now renders at /src
rather than root.

Using `window.location.href` works so long as both the login
button/link/what-have-you is located at a route that can
also figure out what to do with an authed user - which, in
this case, it is for both the "complete" version as well as
the chapter versions.